pyproject.toml
[project]
name = "andrey"
version = "0.1.0"
description = "AI-powered Android game testing assistant"
requires-python = ">=3.10"
dependencies = [
    "adbutils>=2.0",
    "anthropic>=0.40",
    "click>=8.0",
    "pydantic>=2.0",
    "Pillow>=10.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock",
]
omniparser = [
    "torch>=2.0",
    "torchvision",
    "ultralytics>=8.3",
    "transformers>=4.38,<5",
    "easyocr",
    "supervision==0.18.0",
    "accelerate",
    "timm",
    "einops==0.8.0",
]

[project.scripts]
andrey = "andrey.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
